# gotest/docs/swagger.yaml
definitions:
  rsp.ErrorResponse:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 附加数据，通常为额外的错误信息
      message:
        description: 错误描述
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 用户通过用户名和密码登录，返回 JWT token
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回 JWT token
          schema:
            type: string
        "400":
          description: 用户名或密码错误
          schema:
            type: string
        "401":
          description: 用户未授权
          schema:
            type: string
      summary: 用户登录接口
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: 用户通过用户名和密码注册
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功，返回 JWT token
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      summary: 用户注册接口
      tags:
      - Auth
  /articles/{id}/like:
    post:
      description: 根据文章ID，为指定文章增加一次点赞数。
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: 点赞文章
      tags:
      - 文章操作
  /articles/{id}/likes:
    get:
      description: 根据文章ID，获取指定文章的点赞数。
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: 获取文章点赞数
      tags:
      - 文章操作
  /teams:
    post:
      description: 创建一个新的团队
      parameters:
      - description: 团队名称
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: 团队描述
        in: body
        name: description
        schema:
          type: string
      - description: 团队创建者的用户ID
        in: body
        name: owner_id
        required: true
        schema:
          type: integer
      responses:
        "201":
          description: 注册成功，返回 JWT token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rsp.ErrorResponse'
      summary: 新建团队
      tags:
      - 团队操作
  /teams/{id}/members:
    get:
      description: 查询指定团队的所有成员信息
      parameters:
      - description: 团队ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 注册成功，返回 JWT token
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rsp.ErrorResponse'
      summary: 查询团队成员
      tags:
      - 团队操作
  /teams/add_user:
    post:
      description: 将指定用户添加到团队中，并为其设置角色和权限
      parameters:
      - description: 团队ID
        in: body
        name: team_id
        required: true
        schema:
          type: integer
      - description: 用户ID
        in: body
        name: user_id
        required: true
        schema:
          type: integer
      - description: 用户角色（owner/admin/member）
        in: body
        name: role
        required: true
        schema:
          type: string
      - description: 权限（JSON格式）
        in: body
        name: permissions
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 注册成功，返回 JWT token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rsp.ErrorResponse'
      summary: 添加用户到团队
      tags:
      - 团队操作
swagger: "2.0"